requirements.txt
streamlit
pandas
numpy
sentence-transformers
faiss-cpu
scikit-learn
requests
google-generativeai